% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_epa.R
\name{install_epa}
\alias{install_epa}
\title{Download EPA models}
\usage{
install_epa(model, dir = getwd())
}
\arguments{
\item{model}{Names of EPA models to download. Supply multiple models as a vector - ex. 'c("bpip", "aermod")'.
Options: aermap, aermod, aerscreen, bpip, makemet.}

\item{dir}{Folder for installation. Defaults to working directory.}
}
\description{
Download EPA models and install to entered folder.
}
\examples{
install_epa(model = "aerscreen, bpip", dir = "bpip_exe")
}
\keyword{EPA}
\keyword{download}
\keyword{install}
\keyword{models}

