% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installEPA-package.r
\docType{package}
\name{installEPA}
\alias{installEPA}
\alias{installEPA-package}
\title{installEPA: Download EPA air models}
\description{
Download a list of EPA air models. Options: aermap, aermod, aerscreen, bpip, makemet.
}

