% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_epa.R
\name{install_epa}
\alias{install_epa}
\title{Download EPA air models}
\usage{
install_epa(model, dir = getwd())
}
\arguments{
\item{model}{Names of EPA models to download. Supply multiple models as a vector -> \code{c("bpip", "aermod")}.
Options: aermap, aermet, aermod, aerscreen, bpip, makemet.}

\item{dir}{Folder for installation. Defaults to working directory.}
}
\description{
Download the most recent version of EPA air models and install to chosen directory.
}
\examples{
install_epa(model = c("aerscreen", "bpip"), dir = "bpip_exe")
}
\keyword{EPA}
\keyword{download}
\keyword{install}
\keyword{models}

